spring: correccion de excepciones

+++ src/main/java/com/nxhu/restfull_api/product/infraestructure/adapter/ouput/persistence/entity/ProductEntity.java
@@ -26,5 +26,5 @@ public class ProductEntity {
 	
 	private String name;
 
-	private float price;
+	private double price;
 }
